separeted chart concerns
