+++ src/views/LHD/components/HistoricalChart/index.tsx
@@ -8,23 +8,14 @@ import { Line } from 'react-chartjs-2'
 import { Flex } from 'components/uikit'
 
 // Helpers
-import { parseHistoricalData } from './parseHistoricalData'
+import { getChartOptions } from './utils/getChartOptions'
+import { getDataSets } from './utils/getDataSets'
 
 // Types
 import { HistoricTokenData } from 'state/lhd/types'
+import { DataSetNames } from './types'
 
 ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, Tooltip)
-
-enum dataSetNames {
-  dataSetOne = 'dataSetOne',
-  dataSetTwo = 'dataSetTwo',
-  dataSetThree = 'dataSetThree',
-  dataSetFour = 'dataSetFour',
-  dataSetFive = 'dataSetFive',
-  dataSetSix = 'dataSetSix',
-  dataSetSeven = 'dataSetSeven',
-  dataSetEight = 'dataSetEight',
-}
 
 const HistoricalChart = ({
   tokenHistoric,
@@ -33,227 +24,28 @@ const HistoricalChart = ({
   tokenHistoric: HistoricTokenData[] | never[]
   isLoading: boolean
 }) => {
-  const {
-    mcap,
-    ownershipScore,
-    concentrationScore,
-    totalScore,
-    totalExtractableLiquidity,
-    healthScore,
-    ownedLiquidity,
-    liquidityDebt,
-  } = useMemo(() => parseHistoricalData(tokenHistoric), [tokenHistoric])
+  const data = useMemo(() => getDataSets(tokenHistoric), [tokenHistoric])
 
-  const dataSetOne = {
-    label: dataSetNames.dataSetOne,
-    data: mcap,
-    borderColor: 'rgb(255, 99, 132)',
-    backgroundColor: 'rgba(255, 99, 132, 1)',
-    yAxisID: 'y',
-  }
-
-  const dataSetTwo = {
-    label: dataSetNames.dataSetTwo,
-    data: ownershipScore,
-    borderColor: 'rgb(53, 162, 235)',
-    backgroundColor: 'rgba(53, 162, 235, 0.5)',
-    yAxisID: 'y1',
-  }
-
-  const dataSetThree = {
-    label: dataSetNames.dataSetThree,
-    data: concentrationScore,
-    borderColor: 'rgb(124, 252, 0)',
-    backgroundColor: 'rgba(124, 252, 0, 1)',
-    yAxisID: 'y2',
-  }
-
-  const dataSetFour = {
-    label: dataSetNames.dataSetFour,
-    data: totalExtractableLiquidity,
-    borderColor: 'rgb(255, 192, 203)',
-    backgroundColor: 'rgba(255, 192, 203, 1)',
-    yAxisID: 'y3',
-  }
-
-  const dataSetFive = {
-    label: dataSetNames.dataSetFive,
-    data: totalScore,
-    borderColor: 'rgb(238, 75, 43)',
-    backgroundColor: 'rgba(238, 75, 43, 1)',
-    yAxisID: 'y4',
-  }
-
-  const dataSetSix = {
-    label: dataSetNames.dataSetSix,
-    data: healthScore,
-    borderColor: 'rgb(255, 172, 28)',
-    backgroundColor: 'rgba(255, 172, 28, 1)',
-    yAxisID: 'y5',
-  }
-
-  const dataSetSeven = {
-    label: dataSetNames.dataSetSeven,
-    data: ownedLiquidity,
-    borderColor: 'rgb(230, 230, 250)',
-    backgroundColor: 'rgba(230, 230, 250, 1)',
-    yAxisID: 'y6',
-  }
-
-  const dataSetEight = {
-    label: dataSetNames.dataSetEight,
-    data: liquidityDebt,
-    borderColor: 'rgb(139,69,19)',
-    backgroundColor: 'rgba(139, 69, 19, 1)',
-    yAxisID: 'y6',
-  }
-
-  const datasets = useMemo(
-    () => [dataSetOne, dataSetTwo, dataSetThree, dataSetFour, dataSetFive, dataSetSix, dataSetSeven, dataSetEight],
-    [dataSetOne, dataSetTwo, dataSetThree, dataSetFour, dataSetFive, dataSetSix, dataSetSeven, dataSetEight],
-  )
-
-  const data = useMemo(() => {
-    return {
-      labels: Array(mcap.length).fill(''),
-      datasets: datasets,
-    }
-  }, [tokenHistoric])
   const [chartData, setChartData] = useState(data)
   const [toggledData, setToggledData] = useState({
-    [dataSetNames.dataSetOne]: true,
-    [dataSetNames.dataSetTwo]: true,
-    [dataSetNames.dataSetThree]: true,
-    [dataSetNames.dataSetFour]: true,
-    [dataSetNames.dataSetFive]: true,
-    [dataSetNames.dataSetSix]: true,
-    [dataSetNames.dataSetSeven]: true,
-    [dataSetNames.dataSetEight]: true,
+    [DataSetNames.dataSetOne]: true,
+    [DataSetNames.dataSetTwo]: true,
+    [DataSetNames.dataSetThree]: true,
+    [DataSetNames.dataSetFour]: true,
+    [DataSetNames.dataSetFive]: true,
+    [DataSetNames.dataSetSix]: true,
+    [DataSetNames.dataSetSeven]: true,
+    [DataSetNames.dataSetEight]: true,
   })
 
   useEffect(() => {
     setChartData(data)
   }, [data])
 
-  const options = {
-    responsive: true,
-    interaction: {
-      mode: 'index' as const,
-      intersect: false,
-    },
-    elements: {
-      point: {
-        radius: 0,
-      },
-    },
-    stacked: false,
-    plugins: {
-      title: {
-        display: true,
-        text: 'Chart.js Line Chart - Multi Axis',
-      },
-    },
-    scales: {
-      y: {
-        type: 'linear' as const,
-        // display: toggledData[dataSetNames.dataSetOne],
-        display: false,
-        position: 'left' as const,
-        title: {
-          display: true,
-          text: 'One',
-        },
-      },
-      y1: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'two',
-        },
-      },
-      y2: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Three',
-        },
-      },
-      y3: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Four',
-        },
-      },
-      y4: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Five',
-        },
-      },
-      y5: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Six',
-        },
-      },
-      y6: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Seven',
-        },
-      },
-      y7: {
-        type: 'linear' as const,
-        display: false,
-        position: 'left' as const,
-        grid: {
-          drawOnChartArea: false,
-        },
-        title: {
-          display: false,
-          text: 'Eight',
-        },
-      },
-    },
-  }
+  const options = getChartOptions()
 
-  const handleDataToggle = ({ dataSetName }: { dataSetName: dataSetNames }) => {
+  const handleDataToggle = ({ dataSetName }: { dataSetName: DataSetNames }) => {
     setToggledData({ ...toggledData, [dataSetName]: !toggledData[dataSetName] })
-
     const newChartData = data.datasets.filter((set) => {
       if (set.label === dataSetName) {
         return !toggledData[dataSetName]
@@ -275,14 +67,14 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(255, 99, 132)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetOne })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetOne })}
         >
           One Mcap
         </button>
         <button
           style={{ backgroundColor: 'rgb(53, 162, 235)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetTwo })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetTwo })}
         >
           Two Ownership
         </button>
@@ -290,14 +82,14 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(124, 252, 0)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetThree })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetThree })}
         >
           Three Concentration Score
         </button>
         <button
           style={{ backgroundColor: 'rgb(255, 192, 203)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetFour })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetFour })}
         >
           Four Total Extractable
         </button>
@@ -305,7 +97,7 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(238, 75, 43)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetFive })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetFive })}
         >
           Five Total Score
         </button>
@@ -313,7 +105,7 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(255, 172, 28)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetSix })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetSix })}
         >
           Six Health Score
         </button>
@@ -321,7 +113,7 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(230, 230, 250)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetSeven })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetSeven })}
         >
           Seven Owned Liquidity
         </button>
@@ -329,7 +121,7 @@ const HistoricalChart = ({
         <button
           style={{ backgroundColor: 'rgb(139,69,19)' }}
           type="button"
-          onClick={() => handleDataToggle({ dataSetName: dataSetNames.dataSetEight })}
+          onClick={() => handleDataToggle({ dataSetName: DataSetNames.dataSetEight })}
         >
           Eight Liquidity Debt
         </button>
