+++ src/views/LHD/components/HistoricalChart/types.ts
@@ -1,0 +1,10 @@
+export enum DataSetNames {
+  dataSetOne = 'dataSetOne',
+  dataSetTwo = 'dataSetTwo',
+  dataSetThree = 'dataSetThree',
+  dataSetFour = 'dataSetFour',
+  dataSetFive = 'dataSetFive',
+  dataSetSix = 'dataSetSix',
+  dataSetSeven = 'dataSetSeven',
+  dataSetEight = 'dataSetEight',
+}
