+++ src/views/LHD/components/HistoricalChart/utils/parseHistoricalData.ts
@@ -1,3 +1,4 @@
+// types
 import { HistoricTokenData } from 'state/lhd/types'
 
 interface IParsedHistoricalData {
